requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SocketIO==5.1.1
geoip2==4.2.0
mysqlclient==2.0.3
python-dotenv==0.19.0
Werkzeug==2.0.1 